% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{quiverPlot}
\alias{quiverPlot}
\title{Plot of velocity field}
\usage{
quiverPlot(frameData, arrowl, colormapped = F)
}
\arguments{
\item{frameData}{A 4-column data frame consisting of (from left to right) the x position, y position, x-component, and y-component of the velocity.}

\item{arrowl}{A scalar multiple to apply to the velocity components, for visualization.}

\item{colormapped}{If F, plot as arrow line segments. Else, plot as color heatmap.}
}
\value{
A ggplot2 object that is the plot of the velocity field.
}
\description{
This function takes in a 4-column data frame that contains the requisite data on a velocity field, where each row is
a vector in space which has an x and y position and an x and y velocity component. It returns a ggplot object.
}
