% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{readLongformAtChunks}
\alias{readLongformAtChunks}
\title{Read specified chunks of data in an HDF5 dataset.}
\usage{
readLongformAtChunks(myFileName, datasetName, chunkDefineTable)
}
\arguments{
\item{myFileName}{The HDF5 file from which to load the data}

\item{datasetName}{The dataset name within the HDF5 file}

\item{chunkDefineTable}{A table that contains a dataChunkStart and dataChunkEnd column.}
}
\description{
This function reads chunks from an HDF5 file with the name provided in filename and the chunks delimited by a data
 frame called chunkDefineTable. This function should in general not be called, except through the readDataAtFrame function.
}
\details{
chunkDefineTable must contain two columns with the names dataChunkStart and dataChunkEnd that specify the blocks
of data to load.
}
