% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{reshapeArrayLongForm}
\alias{reshapeArrayLongForm}
\title{Transform matrix format to flatfile format.}
\usage{
reshapeArrayLongForm(myarr)
}
\arguments{
\item{myarr}{A N-dimensional matrix.}
}
\value{
A data frame with (X*Y*Z...) columns rows and N columns.
}
\description{
Transforms a matrix of values into a long-form data frame.
}
